Codebase file restructure worked well
