+++ src/app/blog/layout.tsx
@@ -1,5 +1,6 @@
 import type { Metadata } from 'next';
-import { BlogContent } from '../../../components/BlogContent';
+
+import { BlogContent } from '@/components/blog/blog-content';
 
 export const metadata: Metadata = {
   title: 'Blog | Hugo Bayoud Développeur Full Stack Web & Mobile',
