+++ src/components/ui/theme-toggle.tsx
@@ -2,8 +2,9 @@
 
 import { Button } from '@radix-ui/themes';
 import { MoonIcon, SunIcon } from '@radix-ui/react-icons';
-import { useTheme } from '../providers/theme-provider';
-import { useLanguage } from '../src/i18n/LanguageProvider';
+
+import { useTheme } from '@/lib/hooks/use-theme';
+import { useLanguage } from '@/lib/hooks/use-language';
 
 export function ThemeToggle() {
   const { theme, toggleTheme } = useTheme();
